Add files via upload
Im not very proud of this as I got kind of frustrated trying to meet the conditions. But I have learned a lot since completing this project and I will probably do it again. I will still upload it as im not afraid of showing my failures because this is how we learn!